Extract SelectedLocationHits component
